% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dennis_gam_initial_functions.R
\name{abundance_index}
\alias{abundance_index}
\title{abundance_index Function}
\usage{
abundance_index(your_dataset, flight_pheno)
}
\arguments{
\item{your_dataset}{A data.frame containing original species count data. The data format is a csv (comma "," separated) with 6 columns with headers, namely    "species","transect_id","visit_year","visit_month","visit_day","sp_count"}

\item{flight_pheno}{A data.frame for the regional flight curve computed with the function flight_curve}
}
\description{
This function compute the Abundance Index across sites and years from your dataset and the regional flight curve
}
\examples{
abundance_index()
}
\keyword{curve}
\keyword{flight}
\keyword{standardize}

