% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dennis_gam_initial_functions.R
\name{flight_curve}
\alias{flight_curve}
\title{flight_curve Function
This function compute the flight curve across and years}
\usage{
flight_curve(your_dataset)
}
\arguments{
\item{your_dataset}{A data.frame containing original species count data. The data format is a csv (comma "," separated) with 6 columns with headers, namely "species","transect_id","visit_year","visit_month","visit_day","sp_count"}
}
\description{
flight_curve Function
This function compute the flight curve across and years
}
\examples{
flight_curve()
}
\keyword{curve}
\keyword{flight}
\keyword{standardize}

