% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dennis_gam_initial_functions.R
\name{trap_area}
\alias{trap_area}
\title{trap_area Function}
\usage{
trap_area(x, y = NULL)
}
\arguments{
\item{x}{A vector or a two columns matrix.}

\item{y}{A vector, Default is NULL}
}
\description{
This function compute the area under the curve using the trapezoid method.
}
\examples{
trap_area()
}
\keyword{trapezoid}

