% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dennis_gam_initial_functions.R
\name{trap_index}
\alias{trap_index}
\title{trap_index Function}
\usage{
trap_index(sp_data, data_col = "IMP", time_col = "DAYNO",
  by_col = c("SPECIES", "SITE", "YEAR"))
}
\arguments{
\item{sp_data}{A data.frame containing species count data generated from the year_day_func()}

\item{y}{A vector, Default is NULL}
}
\description{
This function compute the area under the curve (Abundance Index) across species, sites and years
}
\examples{
trap_index()
}
\keyword{Abundance}
\keyword{index}

