% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dennis_gam_initial_functions.R
\name{year_day_func}
\alias{year_day_func}
\title{year_day_func Function
This function generate the full sequence of days, months and include the observation to that file.}
\usage{
year_day_func(sp_data)
}
\arguments{
\item{sp_data}{A data.frame with your observation.}
}
\description{
year_day_func Function
This function generate the full sequence of days, months and include the observation to that file.
}
\examples{
year_day_func()
}
\author{
Reto Schmucki
}
\keyword{days}
\keyword{year}
